Create README.md
look I'm not totally computer illiterate but until last year I could barely open or move a file. I am a decently intelligent person and am trying to change my ways, and what I mean by that is the lesser of two evils. exploring the areas of finance in the shade, so I have utilized certain party favors to stay awake and focus and have crash coursed my self on the subject. Now we all know there is a lot to the cc game and it involves some pretty complicated shit. Anything and everything can be learned and even mastered by the right ones. I'm ok with just being ok or even good at it. Im close to being good but need help with antidetect install and set up and actuall use